Factor normal vector helpers from BoundaryCorrections
We need to do the same manipulations for the inverse spatial metric and the
normal vectors in the generic boundary condition test.
